objectives/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
openpyxl==3.1.2
odfpy==1.4.1
openai==1.6.1
python-dotenv==1.0.0
matplotlib==3.8.2
numpy<2
